Au revoir !         %  & numberOfContigs & bases & meanSize  & n50  & max   & coverage   & misassembled & mismatches & indels  Ray & 128 & 2003745 & 15654 & 28294 &  89893 &  0.9829 & 0 & 1 & 0 \\ [boiseb01@ls30 Open-MPI]$ bash strept-error-ompi.sh |tee log

+++ code/SeedingData.cpp
@@ -67,7 +67,7 @@ void SeedingData::computeSeeds(){
 		}else if(!m_splayTreeIterator.hasNext()){
 			m_currentTreeIndex++;
 			#ifdef ASSERT
-			assert(m_SEEDING_i<m_subgraph->size());
+			assert(m_SEEDING_i<(int)m_subgraph->size());
 			assert(m_currentTreeIndex<m_subgraph->getNumberOfTrees());
 			#endif
